# Update to Webpack 3. Enable ModuleConcatenation