--- a/package.json
+++ b/package.json
@@ -155,7 +155,7 @@
     "node-gyp": "3.6.2",
     "redux-logger": "3.0.6",
     "spectron": "3.6.4",
-    "webpack": "2.6.1",
+    "webpack": "^3.0.0",
     "xo": "0.17.1"
   }
 }