I build locally a version of hyper with Webpack 3 and scope hoisting in the build. I haven't spotted any bugs on Mac (though it would be good if someone on Linux and Windows could generate a build from this branch and test).

Overall, this version started faster than the current one.

I started the app several times from scratch, here are some results:

**without** scope hoisting:
- 291.74ms
- 282.86ms
- 279.19ms

**with** scope hoisting:
- 225.94ms
- 230.86ms
- 254.24ms

I'm testing in a MacBook Air and the difference can be noticeable. Of course this PR doesn't need to be merged, the change is easy and you can do it yourselves, but just as a reminder for the next release in case you want to consider it.