@Stanzilla 2017-07-03
Is tree shaking enabled by default in @3 now? If not, might be worth doing that as well http://jakewiesler.com/tree-shaking-es6-modules-in-webpack-2/