@jeremenichelli 2017-07-03
@Stanzilla seeing that we are not transpiling modules on babel, Webpack does tree shaking by default on _harmony modules_ and dead code elimination should get rid of unused functions already. Further reading: http://2ality.com/2015/12/webpack-tree-shaking.html