@ppot 2017-08-08
Thanks for the PR, I will close since it's integrated into #1958